UDP packet 调试

+++ layers/AppLayer.h
@@ -20,18 +20,18 @@ public:
 
 	void handleSend(Block *block) override;
 
-	void handleUDPData(unsigned char *data, int length);
+	void handleUDPData(unsigned char *data, long long length);
 
 	void start() override;
 
 	void stop() override;
 
-	void sendUDPData(const IP& ip, unsigned char *data, int length);
+	void sendUDPData(const IP& ip, unsigned char *data, long long length);
 
 private:
 	void sendUDPData0(const IP& ip, unsigned char *data, int len);
 	std::thread * thread = nullptr;
-	code_machina::BlockingQueue<std::pair<unsigned char*, std::pair<IP,int>>*> queue;
+	code_machina::BlockingQueue<std::pair<unsigned char*, std::pair<IP,long long>>*> queue;
 	UDPPreTable table = UDPPreTable(this);
 	UDPTable udpTable = UDPTable(this);
 	bool shouldThreadStop = false;
